+++ rust/operator/src/lib.rs
@@ -56,6 +56,10 @@ use strum::IntoEnumIterator;
 use tracing::error;
 use tracing::{debug, info, trace, warn};
 
+/// The docker image we default to. This needs to be adapted if the operator does not work
+/// with images 0.0.1, 0.1.0 etc. anymore and requires e.g. a new major version like 1(.0.0).
+const DEFAULT_IMAGE_VERSION: &str = "0";
+
 const FINALIZER_NAME: &str = "hbase.stackable.tech/cleanup";
 const ID_LABEL: &str = "hbase.stackable.tech/id";
 const SHOULD_BE_SCRAPED: &str = "monitoring.stackable.tech/should_be_scraped";
@@ -435,8 +439,9 @@ impl HbaseState {
         let mut container_builder = ContainerBuilder::new(APP_NAME);
         container_builder.image(format!(
             // TODO: how to handle the platform version?
-            "docker.stackable.tech/stackable/hbase:{}-0.1",
-            version.to_string()
+            "docker.stackable.tech/stackable/hbase:{}-stackable{}",
+            version.to_string(),
+            DEFAULT_IMAGE_VERSION
         ));
         container_builder.command(HbaseRole::from_str(pod_id.role())?.command());