+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - `operator-rs` `0.3.0` → `0.4.0` ([#25]).
+- `stackable-hdfs-crd` `0.1.0` → `0.2.0` ([#25]).
+- `stackable-zookeeper-crd` `0.4.1` → `0.5.0` ([#25]).
 - Adapted pod image and container command to docker image ([#25]).
 - Adapted documentation to represent new workflow with docker images ([#25]).